Fix mocha test server startup
